etl/postgres_to_es/django_app/requirements.txt
Django==4.2
elasticsearch==7.17.7
django-elasticsearch-dsl==7.4.0
psycopg2-binary==2.9.6
backoff==2.2.1
python-dotenv==1.0.0
